fixed from PVS-Studio:
V668 There is no sense in testing the 'junc->emitter' pointer against null, as the memory was allocated using the 'new' operator. The exception will be generated in the case of memory allocation error. emitter.cpp 41
V668 There is no sense in testing the 'node->qualSource' pointer against null, as the memory was allocated using the 'new' operator. The exception will be generated in the case of memory allocation error. qualsource.cpp 41