fixed from PVS-Studio:
V668 There is no sense in testing the 'memBlock->block' pointer against null, as the memory was allocated using the 'new' operator. The exception will be generated in the case of memory allocation error. mempool.cpp
V668 There is no sense in testing the 'memBlock' pointer against null, as the memory was allocated using the 'new' operator. The exception will be generated in the case of memory allocation error. mempool.cpp